+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Updated `profile.ps1` to change `$env:PSModulePath` to not use OneDrive folders
 
 ## [3.1.2] - 2021-12-08
 ### Fixed